Faturalar Modülü Güncelleme
Fatura detay kısmında firma müşteri ayrımı yapıldı.
Firma, Müşteri ve Personel isimlerini 'lookupedit' aracı kullanılarak çekildi.
Genel düzeltmeler yapıldı.

+++ Ticari_Otamasyon/Ticari_Otamasyon/FrmFaturalar.cs
@@ -37,39 +37,124 @@ namespace Ticari_Otamasyon
             mskSaat.Text = "";
             txtVergiDaire.Text = "";
             txtAlici.Text = "";
-            txtTeslimEden.Text = "";
-            txtTeslimAlan.Text = "";
 
-            txtUrunID.Text = "";
             txtUrunAd.Text = "";
             txtMiktar.Text = "";
             txtFiyat.Text = "";
+            txtMarka.Text = "";
             txtTutar.Text = "";
             txtFaturaID.Text = "";
+            cmbPersonel.Text = "";
+            cmbPersonel2.Text = "";
+            comboBox1.SelectedItem = null;
+            lookUpEdit1.EditValue = null;
+            lookUpEdit2.EditValue = null;
+        }
 
+        void PersonelListesi()
+        {
+            DataTable dt = new DataTable();
+            SqlDataAdapter da = new SqlDataAdapter("SELECT ID,(AD+' '+SOYAD) AS 'İSİM' FROM TBL_PERSONELLER", bgl.baglanti());
+            da.Fill(dt);
+            cmbPersonel.ValueMember = "ID";
+            cmbPersonel.DisplayMember = "İSİM";
+            cmbPersonel.DataSource = dt;
+
+            DataTable dt1 = new DataTable();
+            SqlDataAdapter da1 = new SqlDataAdapter("SELECT ID,(AD+' '+SOYAD) AS 'İSİM' FROM TBL_PERSONELLER", bgl.baglanti());
+            da1.Fill(dt1);
+
+            cmbPersonel2.ValueMember = "ID";
+            cmbPersonel2.DisplayMember = "İSİM";
+            cmbPersonel2.DataSource = dt1;
+
+            bgl.baglanti().Close();
+        }
+
+        void FaturaDetay()
+        {
+            double miktar, fiyat, tutar;
+            fiyat = Convert.ToDouble(txtFiyat.Text);
+            miktar = Convert.ToDouble(txtMiktar.Text);
+            tutar = fiyat * miktar;
+            txtTutar.Text = tutar.ToString();
+            SqlCommand save = new SqlCommand("INSERT INTO TBL_FATURADETAY (URUNAD,MIKTAR,FIYAT,TUTAR,FATURAID,MARKA) VALUES (@s1,@s2,@s3,@s4,@s5,@s6)", bgl.baglanti());
+            save.Parameters.AddWithValue("@s1", txtUrunAd.Text);
+            save.Parameters.AddWithValue("@s2", txtMiktar.Text);
+            save.Parameters.AddWithValue("@s3", decimal.Parse(txtFiyat.Text));
+            save.Parameters.AddWithValue("@s4", decimal.Parse(txtTutar.Text));
+            save.Parameters.AddWithValue("@s5", txtFaturaID.Text);
+            save.Parameters.AddWithValue("@s6", txtMarka.Text);
+            save.ExecuteNonQuery();
+            bgl.baglanti().Close();
+        }
+
+        void UrunListesi()
+        {
+            DataTable dt = new DataTable();
+            SqlDataAdapter da = new SqlDataAdapter("SELECT ID,URUNAD,MARKA FROM TBL_URUNLER WHERE ADET>0 ", bgl.baglanti());
+            da.Fill(dt);
+            lookUpEdit1.Properties.ValueMember = "ID";
+            lookUpEdit1.Properties.DisplayMember = "ID";
+            lookUpEdit1.Properties.DataSource = dt;
+        }
+
+        void FirmaListesi()
+        {
+            DataTable dt = new DataTable();
+            SqlDataAdapter da = new SqlDataAdapter("SELECT ID,AD FROM TBL_FIRMALAR", bgl.baglanti());
+            da.Fill(dt);
+
+            lookUpEdit2.Properties.ValueMember = "ID";
+            lookUpEdit2.Properties.DisplayMember = "AD";
+            lookUpEdit2.Properties.DataSource = dt;
+        }
+
+        void MusteriListesi()
+        {
+            DataTable dt = new DataTable();
+            SqlDataAdapter da = new SqlDataAdapter("SELECT ID,(AD+' '+SOYAD) AS 'İSİM' FROM TBL_MUSTERILER", bgl.baglanti());
+            da.Fill(dt);
+            lookUpEdit2.Properties.ValueMember = "ID";
+            lookUpEdit2.Properties.DisplayMember = "İSİM";
+            lookUpEdit2.Properties.DataSource = dt;
         }
 
         private void FrmFaturalar_Load(object sender, EventArgs e)
         {
             FaturaListesi();
+            UrunListesi();
+            PersonelListesi();
             Temizle();
+        }
+
+        private void btnBul_Click(object sender, EventArgs e)
+        {
+            SqlCommand komut = new SqlCommand("SELECT URUNAD,MARKA,SATISFIYAT FROM TBL_URUNLER WHERE ID=@p1", bgl.baglanti());
+            komut.Parameters.AddWithValue("@p1", lookUpEdit1.EditValue);
+            SqlDataReader dr = komut.ExecuteReader();
+            while (dr.Read())
+            {
+                txtUrunAd.Text = dr[0].ToString();
+                txtMarka.Text = dr[1].ToString();
+                txtFiyat.Text = dr[2].ToString();
+            }
+            bgl.baglanti().Close();
         }
 
         private void btnKaydet_Click(object sender, EventArgs e)
         {
-           
-            if (txtFaturaID.Text == "" )
+            if (txtFaturaID.Text == "")
             {
-                SqlCommand Kayit = new SqlCommand("INSERT INTO TBL_FATURABILGI (SERI,SIRANO,TARIH,SAAT,VERGIDAIRE,ALICI,TESLIMEDEN,TESLIMALAN) VALUES (@k1,@k2,@k3,@k4,@k5,@k6,@k7,@k8)", bgl.baglanti());
-                Kayit.Parameters.AddWithValue("@k1", txtSeri.Text);
-                Kayit.Parameters.AddWithValue("@k2", txtSeriNo.Text);
-                Kayit.Parameters.AddWithValue("@k3", mskTarih.Text);
-                Kayit.Parameters.AddWithValue("@k4", mskSaat.Text);
-                Kayit.Parameters.AddWithValue("@k5", txtVergiDaire.Text);
-                Kayit.Parameters.AddWithValue("@k6", txtAlici.Text);
-                Kayit.Parameters.AddWithValue("@k7", txtTeslimEden.Text);
-                Kayit.Parameters.AddWithValue("@k8", txtTeslimAlan.Text);
-                Kayit.ExecuteNonQuery();
+                SqlCommand save = new SqlCommand("INSERT INTO TBL_FATURABILGI (SERI,SIRANO,TARIH,SAAT,VERGIDAIRE,ALICI,TESLIMEDEN) VALUES (@k1,@k2,@k3,@k4,@k5,@k6,@k7)", bgl.baglanti());
+                save.Parameters.AddWithValue("@k1", txtSeri.Text);
+                save.Parameters.AddWithValue("@k2", txtSeriNo.Text);
+                save.Parameters.AddWithValue("@k3", mskTarih.Text);
+                save.Parameters.AddWithValue("@k4", mskSaat.Text);
+                save.Parameters.AddWithValue("@k5", txtVergiDaire.Text);
+                save.Parameters.AddWithValue("@k6", txtAlici.Text);
+                save.Parameters.AddWithValue("@k7", cmbPersonel.Text);
+                save.ExecuteNonQuery();
                 bgl.baglanti().Close();
                 MessageBox.Show("Fatura Bilgisi Sisteme Kaydedildi", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
                 FaturaListesi();
@@ -79,41 +164,30 @@ namespace Ticari_Otamasyon
             //Firma Çari
             if (txtFaturaID.Text != "" && comboBox1.Text == "Firma")
             {
-                double miktar, fiyat, tutar;
-                fiyat = Convert.ToDouble(txtFiyat.Text);
-                miktar = Convert.ToDouble(txtMiktar.Text);
-                tutar = fiyat * miktar;
-                txtTutar.Text = tutar.ToString();
-                SqlCommand kayit2 = new SqlCommand("INSERT INTO TBL_FATURADETAY (URUNAD,MIKTAR,FIYAT,TUTAR,FATURAID) VALUES (@s1,@s2,@s3,@s4,@s5)", bgl.baglanti());
-                kayit2.Parameters.AddWithValue("@s1", txtUrunAd.Text);
-                kayit2.Parameters.AddWithValue("@s2", txtMiktar.Text);
-                kayit2.Parameters.AddWithValue("@s3", decimal.Parse(txtFiyat.Text));
-                kayit2.Parameters.AddWithValue("@s4", decimal.Parse(txtTutar.Text));
-                kayit2.Parameters.AddWithValue("@s5", txtFaturaID.Text);
-                kayit2.ExecuteNonQuery();
+                FaturaDetay();
+
+
+                //Satış işlemin frimaHareketler tablosuna ekleme
+                SqlCommand save2 = new SqlCommand("INSERT INTO TBL_FIRMAHAREKETLER (URUNID,ADET,PERSONEL,FIRMA,FIYAT,TOPLAM,FATURAID,MARKA) VALUES (@k1,@k2,@k3,@k4,@k5,@k6,@k7,@k8)", bgl.baglanti());
+                save2.Parameters.AddWithValue("@k1", lookUpEdit1.EditValue);
+                save2.Parameters.AddWithValue("@k2", txtMiktar.Text);
+                save2.Parameters.AddWithValue("@k3", cmbPersonel2.SelectedValue);
+                save2.Parameters.AddWithValue("@k4", lookUpEdit2.EditValue);
+                save2.Parameters.AddWithValue("@k5", decimal.Parse(txtFiyat.Text));
+                save2.Parameters.AddWithValue("@k6", decimal.Parse(txtTutar.Text));
+                save2.Parameters.AddWithValue("@k7", txtFaturaID.Text);
+                save2.Parameters.AddWithValue("@k8", lookUpEdit1.EditValue);
+                save2.ExecuteNonQuery();
                 bgl.baglanti().Close();
 
-                
-                //Satış işlemin frimaHareketler tablosuna ekleme
-                SqlCommand komut = new SqlCommand("INSERT INTO TBL_FIRMAHAREKETLER (URUNID,ADET,PERSONEL,FIRMA,FIYAT,TOPLAM,FATURAID) VALUES (@k1,@k2,@k3,@k4,@k5,@k6,@k7)", bgl.baglanti());
-                komut.Parameters.AddWithValue("@k1", txtUrunID.Text);
-                komut.Parameters.AddWithValue("@k2", txtMiktar.Text);
-                komut.Parameters.AddWithValue("@k3", txtPersonel.Text);
-                komut.Parameters.AddWithValue("@k4", txtFirma.Text);
-                komut.Parameters.AddWithValue("@k5", decimal.Parse(txtFiyat.Text));
-                komut.Parameters.AddWithValue("@k6", decimal.Parse(txtTutar.Text));
-                komut.Parameters.AddWithValue("@k7", txtFaturaID.Text);
-                komut.ExecuteNonQuery();
-                bgl.baglanti().Close();
-                
                 //Stok Azaltma
-                SqlCommand komut2 = new SqlCommand("UPDATE TBL_URUNLER SET ADET=ADET-@p1 WHERE ID=@p2", bgl.baglanti());
-                komut2.Parameters.AddWithValue("@p1",txtMiktar.Text);
-                komut2.Parameters.AddWithValue("@p2", txtUrunID.Text);
-                komut2.ExecuteNonQuery();
+                SqlCommand update = new SqlCommand("UPDATE TBL_URUNLER SET ADET=ADET-@p1 WHERE ID=@p2", bgl.baglanti());
+                update.Parameters.AddWithValue("@p1", txtMiktar.Text);
+                update.Parameters.AddWithValue("@p2", lookUpEdit1.EditValue);
+                update.ExecuteNonQuery();
                 bgl.baglanti().Close();
                 MessageBox.Show("Fatura Ayit Ürün Kaydedildi", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
-               
+
                 FaturaListesi();
                 Temizle();
             }
@@ -121,38 +195,27 @@ namespace Ticari_Otamasyon
             //Müşteri cari
             if (txtFaturaID.Text != "" && comboBox1.Text == "Müşteri")
             {
-                double miktar, fiyat, tutar;
-                fiyat = Convert.ToDouble(txtFiyat.Text);
-                miktar = Convert.ToDouble(txtMiktar.Text);
-                tutar = fiyat * miktar;
-                txtTutar.Text = tutar.ToString();
-                SqlCommand kayit2 = new SqlCommand("INSERT INTO TBL_FATURADETAY (URUNAD,MIKTAR,FIYAT,TUTAR,FATURAID) VALUES (@s1,@s2,@s3,@s4,@s5)", bgl.baglanti());
-                kayit2.Parameters.AddWithValue("@s1", txtUrunAd.Text);
-                kayit2.Parameters.AddWithValue("@s2", txtMiktar.Text);
-                kayit2.Parameters.AddWithValue("@s3", decimal.Parse(txtFiyat.Text));
-                kayit2.Parameters.AddWithValue("@s4", decimal.Parse(txtTutar.Text));
-                kayit2.Parameters.AddWithValue("@s5", txtFaturaID.Text);
-                kayit2.ExecuteNonQuery();
-                bgl.baglanti().Close();
+                FaturaDetay();
 
-                
+
                 //Satış işlemin MusteriHareketler tablosuna ekleme
-                SqlCommand komut = new SqlCommand("INSERT INTO TBL_MUSTERIHAREKETLER (URUNID,ADET,PERSONEL,MUSTERI,FIYAT,TOPLAM,FATURAID) VALUES (@k1,@k2,@k3,@k4,@k5,@k6,@k7)", bgl.baglanti());
-                komut.Parameters.AddWithValue("@k1", txtUrunID.Text);
-                komut.Parameters.AddWithValue("@k2", txtMiktar.Text);
-                komut.Parameters.AddWithValue("@k3", txtPersonel.Text);
-                komut.Parameters.AddWithValue("@k4", txtFirma.Text);
-                komut.Parameters.AddWithValue("@k5", decimal.Parse(txtFiyat.Text));
-                komut.Parameters.AddWithValue("@k6", decimal.Parse(txtTutar.Text));
-                komut.Parameters.AddWithValue("@k7", txtFaturaID.Text);
-                komut.ExecuteNonQuery();
+                SqlCommand save2 = new SqlCommand("INSERT INTO TBL_MUSTERIHAREKETLER (URUNID,ADET,PERSONEL,MUSTERI,FIYAT,TOPLAM,FATURAID,MARKA) VALUES (@k1,@k2,@k3,@k4,@k5,@k6,@k7,@k8)", bgl.baglanti());
+                save2.Parameters.AddWithValue("@k1", lookUpEdit1.EditValue);
+                save2.Parameters.AddWithValue("@k2", txtMiktar.Text);
+                save2.Parameters.AddWithValue("@k3", cmbPersonel2.SelectedValue);
+                save2.Parameters.AddWithValue("@k4", lookUpEdit2.EditValue);
+                save2.Parameters.AddWithValue("@k5", decimal.Parse(txtFiyat.Text));
+                save2.Parameters.AddWithValue("@k6", decimal.Parse(txtTutar.Text));
+                save2.Parameters.AddWithValue("@k7", txtFaturaID.Text);
+                save2.Parameters.AddWithValue("@k8", lookUpEdit1.EditValue);
+                save2.ExecuteNonQuery();
                 bgl.baglanti().Close();
 
                 //Stok Azaltma
-                SqlCommand komut2 = new SqlCommand("UPDATE TBL_URUNLER SET ADET=ADET-@p1 WHERE ID=@p2", bgl.baglanti());
-                komut2.Parameters.AddWithValue("@p1", txtMiktar.Text);
-                komut2.Parameters.AddWithValue("@p2", txtUrunID.Text);
-                komut2.ExecuteNonQuery();
+                SqlCommand update = new SqlCommand("UPDATE TBL_URUNLER SET ADET=ADET-@p1 WHERE ID=@p2", bgl.baglanti());
+                update.Parameters.AddWithValue("@p1", txtMiktar.Text);
+                update.Parameters.AddWithValue("@p2", lookUpEdit1.EditValue);
+                update.ExecuteNonQuery();
                 bgl.baglanti().Close();
                 MessageBox.Show("Fatura Ayit Ürün Kaydedildi", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
 
@@ -171,8 +234,7 @@ namespace Ticari_Otamasyon
             mskSaat.Text = dr["SAAT"].ToString();
             txtVergiDaire.Text = dr["VERGIDAIRE"].ToString();
             txtAlici.Text = dr["ALICI"].ToString();
-            txtTeslimEden.Text = dr["TESLIMEDEN"].ToString();
-            txtTeslimAlan.Text = dr["TESLIMALAN"].ToString();
+            cmbPersonel.Text = dr["TESLIMEDEN"].ToString();
         }
 
         private void btnTemizle_Click(object sender, EventArgs e)
@@ -182,26 +244,48 @@ namespace Ticari_Otamasyon
 
         private void btnSil_Click(object sender, EventArgs e)
         {
-            SqlCommand Delete = new SqlCommand("DELETE FROM TBL_FATURABILGI WHERE FATURABILGIID=@d1", bgl.baglanti());
-            Delete.Parameters.AddWithValue("@d1", txtId.Text);
-            Delete.ExecuteNonQuery();
-            bgl.baglanti().Close();
-            MessageBox.Show("Fatura Silindi", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            try
+            {
+                DialogResult secim = MessageBox.Show("Silmek İstediğinize Eminmisiniz", "Uyarı", MessageBoxButtons.YesNo, MessageBoxIcon.Information);
+                if (secim == DialogResult.Yes)
+                {
+                    SqlCommand Delete = new SqlCommand("DELETE FROM TBL_FATURABILGI WHERE FATURABILGIID=@d1", bgl.baglanti());
+                    Delete.Parameters.AddWithValue("@d1", txtId.Text);
+                    Delete.ExecuteNonQuery();
+                    bgl.baglanti().Close();
+                    MessageBox.Show("Fatura Silindi", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    FaturaListesi();
+                    Temizle();
+                }
+                else
+                {
+                    MessageBox.Show("Silme İşlemi İptal Edilmiştir.", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    FaturaListesi();
+                }
+            }
+            catch
+            {
+
+                MessageBox.Show("Bir Hata Meydana Geldi.Lütfen Silmek İstediğiniz Stüna İki Kere Tıklayarak Tekrar Deneyiniz.!", "UYARI", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
+            finally
+            {
+                Temizle();
+            }
             FaturaListesi();
-            Temizle();
+
         }
 
         private void BtnGuncelle_Click(object sender, EventArgs e)
         {
-            SqlCommand Update = new SqlCommand("UPDATE TBL_FATURABILGI SET SERI=@u1,SIRANO=@u2,TARIH=@u3,SAAT=@u4,VERGIDAIRE=@u5,ALICI=@u6,TESLIMEDEN=@u7,TESLIMALAN=@u8 WHERE FATURABILGIID=@u9", bgl.baglanti());
+            SqlCommand Update = new SqlCommand("UPDATE TBL_FATURABILGI SET SERI=@u1,SIRANO=@u2,TARIH=@u3,SAAT=@u4,VERGIDAIRE=@u5,ALICI=@u6,TESLIMEDEN=@u7 WHERE FATURABILGIID=@u9", bgl.baglanti());
             Update.Parameters.AddWithValue("@u1", txtSeri.Text);
             Update.Parameters.AddWithValue("@u2", txtSeriNo.Text);
             Update.Parameters.AddWithValue("@u3", mskTarih.Text);
             Update.Parameters.AddWithValue("@u4", mskSaat.Text);
             Update.Parameters.AddWithValue("@u5", txtVergiDaire.Text);
             Update.Parameters.AddWithValue("@u6", txtAlici.Text);
-            Update.Parameters.AddWithValue("@u7", txtTeslimEden.Text);
-            Update.Parameters.AddWithValue("@u8", txtTeslimAlan.Text);
+            Update.Parameters.AddWithValue("@u7", cmbPersonel.Text);
             Update.Parameters.AddWithValue("@u9", txtId.Text);
             Update.ExecuteNonQuery();
             bgl.baglanti().Close();
@@ -224,19 +308,16 @@ namespace Ticari_Otamasyon
 
         }
 
-        private void btnBul_Click(object sender, EventArgs e)
+        private void comboBox1_SelectedIndexChanged(object sender, EventArgs e)
         {
-            SqlCommand komut = new SqlCommand("SELECT URUNAD,SATISFIYAT FROM TBL_URUNLER WHERE ID=@p1", bgl.baglanti());
-            komut.Parameters.AddWithValue("@p1", txtUrunID.Text);
-            SqlDataReader dr = komut.ExecuteReader();
-            while (dr.Read())
+            if (comboBox1.Text == "Müşteri")
             {
-                txtUrunAd.Text = dr[0].ToString();
-                txtFiyat.Text = dr[1].ToString();
+                MusteriListesi();
             }
-
-            bgl.baglanti().Close();
+            if (comboBox1.Text == "Firma")
+            {
+                FirmaListesi();
+            }
         }
-  
     }
 }
